refactor core Dockerfile
Using the python:3.10.11-slim-bullseye base image to:
 * be able to build for other platforms in addition to amd64 (important for Apple Silicon).
 * make it possible to have reproducible images

The core application is now copied inside the container image and it is not mounted at runtime in a volume.

From Kubernetes docs:
https://kubernetes.io/docs/concepts/containers/

Containers are intended to be stateless and immutable: you should not change
the code of a container that is already running. If you have a containerized
application and want to make changes, the correct process is to build a new
image that includes the change, then recreate the container to start from the
updated image.

Related docs on the choice of the base image:
https://github.com/tiangolo/uvicorn-gunicorn-fastapi-docker#-warning-you-probably-dont-need-this-docker-image

Creating `appuser` to prevent the application to run as root in the container

Dropping the unused `apt-get update` call that was increasing the container image size.