Refactor the admin container to be immutable
From Kubernetes docs:
https://kubernetes.io/docs/concepts/containers/

Containers are intended to be stateless and immutable: you should not change
the code of a container that is already running. If you have a containerized
application and want to make changes, the correct process is to build a new
image that includes the change, then recreate the container to start from the
updated image.