Speed up build time when only plugins req installation
The dev define a new package in the requirements file.
The dev run the build with `docker build`.
The build recreate all the layers.

If the layer is the same, docker keep it, when a layer changes, docker create the layer and all the layer after it.
Moved the apt-get commands before the others steps.
Moved the commands to copy of plugin and installation of plugins dependencies as last commands.