wip: plugin dependencies check
added a flag `CHECK_PLUGIN_DEPENDENCIES` in the compose
set to `false` by default, that, let's you install plugin dependencies
only during plugin activation.
If it is set to `true` instead, requirements installation is done as
usual.
This flag is helpful for the devs in order to reduce the boot time of
the cat.
the private method `_install_requirements` of Plugin class, now
**filters**
the *requirements.txt* from all the currently installed packages.
NOTE: some packages (i.e. `youtube-transcript-api`) have different package and
module names (package name `youtube-transcript-api` and module name
`youtube_transcript_api`) so in this case the check of being installed
unfortunatelly fails, it is an extreme case.
TODO: find a way to get the module name from package name in order to
solve the issue above