Do not remove the session manually
The handling of the session is already handled centrally.
Closing of a sessions raises the exception WebSocketDisconnect that is already handled in websocket.websocket_endpoint

More, removing the session here from the stray, could hide possible unknown.
For now I suggest to not remove the session, this can help to spots others unknown.