Fix test_deactivate_plugin
Why not using plugin.deactivate()?
the "endpoint" decorator needs the reference to the FastAPI app instance for endpoint removing.
Calling plugin.deactivate() the method CustomEndpoint.deactivate() raises the exception:
"self.cheshire_cat_api is None"

because the method on_finish_plugins_sync_callback is not called so the cheshire_cat_api is not injected